Migration changes for test to run

+++ Database/Migrations/2019_07_25_054721_add_program_id_patient_awv_survey_instance_status_view.php
@@ -13,6 +13,8 @@ class AddProgramIdPatientAwvSurveyInstanceStatusView extends Migration
      */
     public function up()
     {
+        if (isUnitTestingEnv()) {return;}
+    
         $viewName = 'patient_awv_survey_instance_status_view';
         \DB::statement("DROP VIEW IF EXISTS ${viewName}");
         \DB::statement("
